service3/requirements.txt
numpy==1.17.4
Jinja2==2.10.3
MarkupSafe==1.1.1 
Werkzeug==0.16.0 
click==7.0 
flask==1.1.1 
itsdangerous==1.1.0
requests==2.22.0
flask-sqlalchemy==2.4.1
pymysql==0.9.3
